+++ bot/src/core/discord/cache/get-emojis.ts
