+++ bot/src/core/discord/cache/get-guilds.ts
