+++ bot/src/core/discord/commands/_index.ts
@@ -1,4 +1,4 @@
-import { Hello } from '@/commands/hello';
 import { Command } from '@/types/types';
+import { Hello } from './hello';
 
 export const Commands: Command[] = [Hello];
