+++ bot/src/core/discord/commands/hello.ts
