+++ bot/src/helpers/format/format-enrollment-message.ts
@@ -1,12 +1,12 @@
 /* eslint-disable @typescript-eslint/restrict-template-expressions */
 
-import { getEmojis } from '@/cache/getEmojis';
-import { getGuild } from '@/cache/getGuild';
 import { Client } from 'discord.js';
 import { formatDiscordTimestamp } from './format-discord-timestamp';
 import { CURRENT_MTG_FORMAT } from '@/constants/drafty';
 import { ASYNC_EMOJI } from '@/constants/emojis';
 import { getDayONextfWeekTimestamp } from '../dates/get-day-of-next-wee-timestamp';
+import { getEmojis } from '@/core/discord/cache/get-emojis';
+import { getGuild } from '@/core/discord/cache/get-guilds';
 
 export const formatEnrollmentMessage = (client: Client) => {
   const scheduledMessage =
